Move CoreDB type definition into apis dir
