+++ coredb-operator/src/apis/mod.rs
@@ -1,0 +1,1 @@
+pub mod coredb_types;
