+++ coredb-operator/src/controller.rs
@@ -6,7 +6,6 @@ use futures::{
 };
 
 use crate::{
-    defaults,
     psql::{PsqlCommand, PsqlOutput},
     service::reconcile_svc,
     statefulset::{reconcile_sts, stateful_set_from_cdb},
@@ -19,66 +18,24 @@ use kube::{
         events::{Event, EventType, Recorder, Reporter},
         finalizer::{finalizer, Event as Finalizer},
     },
-    CustomResource, Resource,
+    Resource,
 };
 
 use crate::{
-    extensions::reconcile_extensions, postgres_exporter_role::create_postgres_exporter_role,
+    apis::coredb_types::{CoreDB, CoreDBStatus},
+    extensions::{reconcile_extensions, Extension},
+    postgres_exporter_role::create_postgres_exporter_role,
     secret::reconcile_secret,
 };
-use k8s_openapi::{
-    api::core::v1::{Namespace, Pod, ResourceRequirements},
-    apimachinery::pkg::api::resource::Quantity,
-};
+use k8s_openapi::api::core::v1::{Namespace, Pod};
 use kube::runtime::wait::Condition;
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use serde_json::json;
 use std::sync::Arc;
 use tokio::{sync::RwLock, time::Duration};
 use tracing::*;
 
 pub static COREDB_FINALIZER: &str = "coredbs.coredb.io";
-
-/// Generate the Kubernetes wrapper struct `CoreDB` from our Spec and Status struct
-///
-/// This provides a hook for generating the CRD yaml (in crdgen.rs)
-#[derive(CustomResource, Deserialize, Serialize, Clone, Debug, JsonSchema)]
-#[cfg_attr(test, derive(Default))]
-#[kube(kind = "CoreDB", group = "coredb.io", version = "v1alpha1", namespaced)]
-#[kube(status = "CoreDBStatus", shortname = "cdb")]
-#[allow(non_snake_case)]
-pub struct CoreDBSpec {
-    #[serde(default = "defaults::default_replicas")]
-    pub replicas: i32,
-    #[serde(default = "defaults::default_resources")]
-    pub resources: ResourceRequirements,
-    #[serde(default = "defaults::default_storage")]
-    pub storage: Quantity,
-    #[serde(default = "defaults::default_postgres_exporter_enabled")]
-    pub postgresExporterEnabled: bool,
-    #[serde(default = "defaults::default_image")]
-    pub image: String,
-    #[serde(default = "defaults::default_postgres_exporter_image")]
-    pub postgresExporterImage: String,
-    #[serde(default = "defaults::default_port")]
-    pub port: i32,
-    #[serde(default = "defaults::default_uid")]
-    pub uid: i32,
-    #[serde(default = "defaults::default_extensions")]
-    pub extensions: Vec<Extension>,
-    #[serde(default = "defaults::default_stop")]
-    pub stop: bool,
-}
-
-/// The status object of `CoreDB`
-#[derive(Deserialize, Serialize, Clone, Default, Debug, JsonSchema)]
-pub struct CoreDBStatus {
-    pub running: bool,
-    pub extensions: Option<Vec<Extension>>,
-    #[serde(default = "defaults::default_storage")]
-    pub storage: Quantity,
-}
 
 // Context for our reconciler
 #[derive(Clone)]
@@ -89,47 +46,6 @@ pub struct Context {
     pub diagnostics: Arc<RwLock<Diagnostics>>,
     /// Prometheus metrics
     pub metrics: Metrics,
-}
-
-#[derive(Clone, Debug, Deserialize, Eq, Hash, JsonSchema, Serialize, PartialEq)]
-pub struct Extension {
-    pub name: String,
-    #[serde(default = "defaults::default_description")]
-    pub description: String,
-    pub locations: Vec<ExtensionInstallLocation>,
-}
-
-impl Default for Extension {
-    fn default() -> Self {
-        Extension {
-            name: "pg_stat_statements".to_owned(),
-            description: " track planning and execution statistics of all SQL statements executed".to_owned(),
-            locations: vec![ExtensionInstallLocation::default()],
-        }
-    }
-}
-
-
-#[derive(Clone, Debug, Deserialize, Eq, Hash, JsonSchema, Serialize, PartialEq)]
-pub struct ExtensionInstallLocation {
-    pub enabled: bool,
-    // no database or schema when disabled
-    #[serde(default = "defaults::default_database")]
-    pub database: String,
-    #[serde(default = "defaults::default_schema")]
-    pub schema: String,
-    pub version: Option<String>,
-}
-
-impl Default for ExtensionInstallLocation {
-    fn default() -> Self {
-        ExtensionInstallLocation {
-            schema: "public".to_owned(),
-            database: "postgres".to_owned(),
-            enabled: true,
-            version: Some("1.9".to_owned()),
-        }
-    }
 }
 
 #[instrument(skip(ctx, cdb), fields(trace_id))]
