+++ coredb-operator/src/crdgen.rs
@@ -1,4 +1,5 @@
 use kube::CustomResourceExt;
+use controller::apis::coredb_types::CoreDB;
 fn main() {
-    print!("{}", serde_yaml::to_string(&controller::CoreDB::crd()).unwrap())
+    print!("{}", serde_yaml::to_string(&CoreDB::crd()).unwrap())
 }
